* Release 1.0.0 changes (#12)

* [EP-864] Added unit test with stubbed data response (#4)

* Carthage updated with OHHTTPMock

* Added testRequest stubbed

* moved back DeviceDNA

* - Code refactoring and re-indentation


* Small improvements and switch to JPError (#1)

* Renamed the method callback signature from NSError to JPError

* Forgot to update header

* Callback changes (#2)

* Card transactions now store errors and callback once the user cancels the payment

* Finished modifying callback logic for normal transactions

* Ran .clang-format.sh

* Updated unit tests (#3)

* Merged unit tests

* Fixed the Customization input field movement bug (#6)

* Changed the order of the payment methods in the Settings page to match default values (#8)

* Fixed Apple Pay transactions not executing (#7)

* Fix/project renaming (#9)

* - Renaming
- Refactoring

* - Added JP refixes to all NS_ENUM and NS_OPTIONS typedefs

* - Unittests adjustments

* - Clang format

* - Updating example apps

* - Clang format

* - Typo fixed in a method signature
- Constraints fix

* - Unittests disabled `Precompile Bridging Header`

* - Fixed typo in method signature

* Readme URLs update (#11)

* - Update README URLs

* - Added project version number


* Use a 'v' as part of the git tag for releases

* [Release] 1.1.0 (#30)

* Renamed, fixed tests and cleaned up project

* Renamed file headers

* Ran ./clang-format.sh and fixed failing tests

* Added back the 'v' in the podspec

* Attempt to fix build bugs

* Updated .sonarscan path

* Testing some .sonarscan settings

+++ Source/View/SectionView/JPSectionView.m
@@ -1,6 +1,6 @@
 //
 //  JPScrollableSectionView.m
-//  JudoKit-iOS
+//  JudoKit_iOS
 //
 //  Copyright (c) 2020 Alternative Payments Ltd
 //
@@ -154,7 +154,7 @@ static const float kSliderCornerRadius = 10.0F;
 
 - (void)moveSliderToIndex:(NSUInteger)index {
 
-    CGFloat xPosition = kContentPadding + index * self.adaptiveSectionWidth;
+    CGFloat xPosition = kContentPadding + (double)index * self.adaptiveSectionWidth;
     CGFloat yPosition = kContentPadding;
     CGFloat width = self.adaptiveSectionWidth - kContentPadding * 2;
     CGFloat height = kSectionHeight - kContentPadding * 2;
@@ -188,11 +188,11 @@ static const float kSliderCornerRadius = 10.0F;
 - (CGFloat)adaptiveSectionWidth {
     CGFloat screenWidth = UIScreen.mainScreen.bounds.size.width;
 
-    if (self.sections.count * kSectionWidth > screenWidth) {
+    if ((double)self.sections.count * kSectionWidth > screenWidth) {
         return kSectionWidth;
     }
 
-    return (screenWidth - kBackgroundPadding * 2) / self.sections.count;
+    return (screenWidth - kBackgroundPadding * 2) / (double)self.sections.count;
 }
 
 - (UIImageView *)generateSectionImageViewFromImage:(UIImage *)image {
